sync-service: Add missing default for http_api_num_acceptors (#2863)
`http_api_num_acceptors` was added as a configuration option with no
default value which meant that it *had* to be configured or the
`api_server/1` call would raise. This PR adds the missing default (`nil`
so that we don't set a custom value by default) and includes a smoke
test of the default configuration to avoid this happening again.

Also enables the external configuration of the `inspector` and
`registry` components which is needed for current work on `Phoenix.Sync`