fix: buffer SSE messages until up-to-date message (#2869)
This PR addresses the issue raised in
https://github.com/electric-sql/electric/pull/2546#issuecomment-3007137018.
The crux of the issue is that messages might be processed several times
if the SSE connection breaks after processing a message but before
receiving the up-to-date message. This is because we only advance the
offset when processing the up-to-date message, hence, we would fetch
again at the old offset and receive some messages that we already
processed.

The fix is easy: we batch SSE messages until we receive up-to-date and
only process the messages at that point. The hard part was to write a
unit test that reproduces this scenario. After a bit of trial and error
i managed to do that by creating a custom fetch wrapper that proxies the
SSE response but filters out the first up-to-date message and then
forces a refresh of the shapestream at that point.