fix: Ensure `LockConnection` can get properly restart on `ConnMan` restarts (#2866)
While working on https://github.com/electric-sql/stratovolt/issues/699 I
noticed that when the connection manager gets restarted by its
supervisor, when it attempts to start the `LockConnection`, it fails
with `{:already_started, ...}`.

The lock connection process is linked to ConnMan, but it seems like
there's a delay between ConnMan dying and the lock connection (which
under the hood is a `:gen_statem`) is killed, so when ConnMan tries to
start a new lock connection with the same name it runs into this issue,
which causes ConnMan to die and restart and run into the same issue
until the supervisor kills it altogether.

Three suggested solutions:
1. Don't register the lock connection process with a global `name`
1a. We don't use it anywhere, so restarts always manage to start a lock
and the timing issue is sidestepped
1b. The negative is that we now have lock processes not globally
addressed and we might end up trying to hold the lock twice somewhere
(?)
2. Add a retry mechanism, either internally to `LockConnection` or in
`ConnMan`, so that if we get `:already_started` we wait a few
milliseconds and try a few times
2a. I tried this and this works fine, but I don't like the arbitrary
retries and timing attempts
3. Ensure lock is killed before trying to start it
 3a. Strongly dislike this approach - feels forced


I've opted for 1. in this PR, after having tried 2. which worked but I
don't love retries. Happy to hear opinions.


## UPDATE
After discussions with @robacourt (thank you for pointing out that
`terminate` still runs when exiting the same process!), I've made conn
man behave slightly more like a supervisor, following loosely what the
OTP supervisor does but with fewer requirements.

I've added the relevant pids of "children" processes that need to be
shut down before conn man exits - this solves the issue. Happy to remove
the extra processes and only keep the lock connection in there to solve
the immediate bug, but this seems to do the trick.