feat(sync-service): Allow configuration of TCP send timeout (#2881)
AntonOfTheWoods experiences the following error message:
```
Error while streaming response: :timeout
```

This is most probably due to slow response processing on the client side
causing TCP backpressure. AntonOfTheWoods has asked for configuration to
increase the timeout, so this PR introduces `ELECTRIC_TCP_SEND_TIMEOUT`
to allow configuration of the TCP/SSL send timeout.

The default is `30s` as it is right now, but for example the timeout can
be set to 60 seconds with:
```
ELECTRIC_TCP_SEND_TIMEOUT=60s
```