Upgrade Elixir and Erlang/OTP to latest versions (#2857)
Follow-up to https://github.com/electric-sql/electric/pull/2771 where
only the code changes were made to future-proof the codebase for this
version upgrade.

Erlang/OTP 28 was released in May 2025. It's good time to bring our
tooling up to date with the latest upstream releases in anticipation of
Elixir 1.19 release in the coming months.