fix: Add lock connection timeout (#2959)
Fixes https://github.com/electric-sql/electric/issues/2944

We ran into an issue with @icehaunter where the lock connection was
stuck trying to handshake with the DB, providing no visibility on the
fact that it is stuck and thus giving no clues as to why things are not
starting up.

I've added back a timeout of 30 seconds (twice the default of 15sec) -
if connection and handshake take longer than that, the lock connection
exits and our supervision machinery takes over, logging
`:connection_error` stack events, retrying, the works.

Postgrex itself will also log an error that the connection was
handshaking for longer than the specified timeout, which will be a good
way to track how often this happens.