Add the ability to specify a root certificate for the database server (#2952)
To achieve true connection security, encryption must be paired up with
identity verification and this is achieved by validating the server
certificate during the TLS handshake.

Web browsers can get away with using OS-provided bundle of root
certificates or shipping their own bundle to enable pervasive use of
HTTPS. However, some managed Postgres providers use a self-signed root
certificate that's not part of OS bundles, so those need to be
downloaded individually and there needs to be some way to let Electric
know about their existence.

This PR introduces a new configuration option that allows users to
specify a path to the root certificate that will then be used to verify
database server identity when establishing a secure connection to it.

Closes https://github.com/electric-sql/electric/issues/2910.

Corresponding documentation update -
https://github.com/electric-sql/electric/pull/2961