feat: Monitor pooled connections fully and handle failures gracefully (#2969)
Fixes https://github.com/electric-sql/electric/issues/2967

After some extensive research with @alco on why we were seeing `:killed`
messages from the connection pool, we found that it was due to the pool
connections failing to connect (see linked issue description).

I started working entirely within ConnMan but was disgusted at how
monolithic it was becoming so I decided to split the pool logic out to
its own process.

The `Pool` process now handles the actual DbConnection pool, monitors
the pool connections fully by linking them with the "supervisor" process
and doing some bookkeeping.

I keep track of the most recent pool connection error and return that in
case the entire pool dies - this piece of logic can and should become
more fleshed out (perhaps track errors with a timeout?) but I find it to
be _much_ better than what we currently do and it does the trick.

You can reproduce the problem described in the issue by setting PG's
`max_connections` to something lower than 20, which makes the pool
unable to become healthy and thus fail with a `:killed` message.

Through this change, ConnMan now also waits for all the pool connections
to become healthy, and it also handles restarts of the pool with a
backoff, and along with the reporting of the individual pool connection
errors this should sort out Cloud reporting as well.

I intend to add tests + an integration test as well for the case where
we run out of connections and encounter this, but ran out of time. Happy
for someone else to do that or for me to pick this back up when I'm
back.